Add game state. Bugfix.
Fix enemy order issue.
More tower sprite placeholders.
Adjust UI parts.
Adjust Enemy parts.
Fix item Cat CC value issue.

+++ Assets/Scripts/Item/Cat.cs
@@ -15,7 +15,7 @@ namespace CSCI526GameJam {
         [ClassHeader(typeof(Cat))]
 
         [MandatoryFields]
-        [SerializeField] private int bonusTowerCCValue;
+        [SerializeField] private float bonusTowerCCValue;
         #endregion
 
         #region Publics
